% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast.tumors.se.r
\docType{data}
\name{breast.tumors.se}
\alias{breast.tumors.se}
\title{Human Breast Tumors Data}
\source{
The Human Breast Tumors dataset is a companion resource for the
paper of Perou \emph{et al.} (2000), and was downloaded from the Stanford
Genomics Breast Cancer Consortium Portal
\url{http://genome-www.stanford.edu/breast_cancer/molecularportraits/download.shtml}
}
\value{
none
}
\description{
The \code{breast.tumors} data with \code{SummarizedExperiment} class.
}
\references{
Perez-Enciso, M. and Tenenhaus, M. (2003). Prediction of
clinical outcome with microarray data: a partial least squares discriminant
analysis (PLS-DA) approach. \emph{Human Genetics} \bold{112}, 581-592.

Perou, C. M., Sorlie, T., Eisen, M. B., van de Rijn, M., Jeffrey, S. S.,
Rees, C. A., Pollack, J. R., Ross, D. T., Johnsen, H., Akslen, L. A., Fluge,
O., Pergamenschikov, A., Williams, C., Zhu, S. X., Lonning, P. E.,
Borresen-Dale, A. L., Brown, P. O. and Botstein, D. (2000). Molecular
portraits of human breast tumours. \emph{Nature} \bold{406}, 747-752.
}
\keyword{datasets}
