% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrimouse.mae.r
\docType{data}
\name{nutrimouse.mae}
\alias{nutrimouse.mae}
\title{Nutrimouse Dataset}
\source{
The \code{nutrimouse} dataset was provided by Pascal Martin from the
Toxicology and Pharmacology Laboratory, National Institute for Agronomic
Research, French.
}
\value{
none
}
\description{
The \code{nutrimouse} dataset in MultiAssayExperiment class.
}
\examples{
experiments(nutrimouse.mae)
## same as:
nutrimouse.mae@ExperimentList
## sample data
colData(nutrimouse.mae)
## sample map
sampleMap(nutrimouse.mae)
}
\references{
Martin, P. G. P., Guillou, H., Lasserre, F., Déjean, S., Lan,
A., Pascussi, J.-M., San Cristobal, M., Legrand, P., Besse, P. and Pineau,
T. (2007). Novel aspects of PPAR\eqn{\alpha}-mediated regulation of lipid
and xenobiotic metabolism revealed through a multrigenomic study.
\emph{Hepatology} \bold{54}, 767-777.
}
\keyword{datasets}
