% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srbct.se.r
\docType{data}
\name{srbct.se}
\alias{srbct.se}
\title{The SummarizedExperiment format of the \code{srbct} dataset.}
\source{
\url{http://research.nhgri.nih.gov/microarray/Supplement}
}
\value{
none
}
\description{
The SummarizedExperiment format of the \code{srbct} dataset.
}
\references{
Khan et al. (2001). Classification and diagnostic prediction of
cancers using gene expression profiling and artificial neural networks.
\emph{Nature Medicine} \bold{7}, Number 6, June.
}
\keyword{datasets}
