% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linnerud.mae.r
\docType{data}
\name{linnerud.mae}
\alias{linnerud.mae}
\title{Linnerud Dataset}
\format{A \code{MultiAssayExperiment} class containing the following
components: \describe{
\item{assay: exercise)}{matrix with 20 observations on 3 exercise
variables.}
\item{assay: physiological)}{data frame with 20 observations on
3 physiological variables.}
}}
\source{
Tenenhaus, M. (1998), Table 1, page 15.
}
\value{
none
}
\description{
The \code{linnerud} data with \code{MultiAssayExperiment} class.
}
\references{
Tenenhaus, M. (1998). \emph{La regression PLS: theorie et
pratique}. Paris: Editions Technic.
}
\keyword{datasets}
